I created a Login/Register workflow which redirects a user to the Campaigner view.
At the moment you have to manually add module and group entries to the database for this to work. The module name should be the same as the protected Campaigner view. The redirection after login is automatic based on the entries in the database. A user has only one default group so we should consider having a field for whether the users default is the campaigners view or supporters view.